requirements.txt
diffusers==0.27.2
flax==0.8.2
google-generativeai==0.5.2
grpcio-status==1.62.2
importlib_resources==6.4.0
reflex==0.4.8.post1
